refactor: created layout with footer
