+++ src/layout/LayoutWithFooter/styled.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const PageContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+
+  min-height: 100vh;
+`;
