+++ src/pages/NotFoundPage/index.tsx
@@ -2,14 +2,14 @@ import { Link } from 'react-router-dom';
 
 import { Button } from '@/components/UI';
 import { AppRoutes, ICONS } from '@/constants';
-import { Footer } from '@/layout/Footer';
+import { LayoutWithFooter } from '@/layout';
 
-import { NotFoundParagraph, NotFoundSection, NotFoundTitle, PageContainer } from './styles';
+import { NotFoundParagraph, NotFoundSection, NotFoundTitle } from './styled';
 
 const { TwitterIcon } = ICONS;
 
 export const NotFoundPage = () => (
-  <PageContainer>
+  <LayoutWithFooter>
     <NotFoundSection>
       <NotFoundTitle $size="xl6">
         <TwitterIcon width={50} height={50} /> 404
@@ -19,7 +19,5 @@ export const NotFoundPage = () => (
         Go To Main
       </Button>
     </NotFoundSection>
-
-    <Footer />
-  </PageContainer>
+  </LayoutWithFooter>
 );
