+++ src/pages/NotFoundPage/styled.ts
@@ -3,13 +3,6 @@ import styled from 'styled-components';
 import { Paragraph, Title } from '@/components/UI';
 import { flex, media } from '@/styles';
 
-export const PageContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-
-  min-height: 100vh;
-`;
-
 export const NotFoundSection = styled.section`
   flex-grow: 1;
   align-self: center;
